moved CaptionedTextWidget to seperate file
